requirements.txt
requests
lxml
pandas
jupyter
gdal
geopandas
PyQt >= 5.0
xlrd